Seed random threat generator

+++ src/main/java/pl/sszwaczyk/security/threat/Threat.java
@@ -19,5 +19,5 @@ public class Threat {
     List<DatapathId> switches;
 
     private LocalTime startTime;
-    private Long duration; //seconds
+    private Integer duration; //seconds
 }
